microservice-masters-project-mock_api_service/requirements.txt
aiofiles==0.8.0
anyio==3.7.1
asgiref==3.7.2
click==8.1.7
dnspython==2.4.2
email-validator==1.2.1
exceptiongroup==1.1.3
Faker==19.3.1
fastapi==0.75.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.3
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.17.1
typing_extensions==4.7.1
uvicorn==0.17.6
